BUG#969: Partially fixed by at least using the timeout for the individual calls to RunSingleCommand from within the inner ctest instance.  This should be modified to incrementally adjust remaining time.

+++ Source/cmCTest.h
@@ -310,7 +310,7 @@ private:
 
   int                     m_TestModel;
 
-  int                     m_TimeOut;
+  double                  m_TimeOut;
 
   std::string             m_MemoryTester;
   std::string             m_MemoryTesterOptions;
